part4 done except exercise 4.22

+++ part4/controllers/blogs.js
@@ -1,10 +1,15 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
-    response.json(blogs.map(blog => blog.toJSON()))
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+    response.json(blogs)
+    //response.json(blogs.map(blog => blog.toJSON()))
   })
+
 
 blogsRouter.get('/:id', async (request, response) => {
     const blogs = await Blog.findById(request.params.id)
@@ -13,17 +18,62 @@ blogsRouter.get('/:id', async (request, response) => {
     else
       response.status(404).end()
   })
+
   
 blogsRouter.post('/', async (request, response) => {  
-  const blog = new Blog(request.body)
+  const body = request.body
+  console.log(request.token)
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  }) 
+
   const savedBlog = await blog.save()
-  response.json(savedBlog)
-    
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.json(savedBlog.toJSON())
   })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+blogsRouter.put('/:id', async (request, response) => {
+  const body = request.body
+  const blog = {
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes 
+  } 
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new:true})
+  response.json(updatedBlog)
 })
+
+
+blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  
+  const blog = await Blog.findById(request.params.id)
+ 
+  if (blog.user.toString() === decodedToken.id){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  
+}) 
 
 module.exports = blogsRouter
